+++ Inc/ft_irc.hpp
@@ -5,12 +5,45 @@
 #include "define.hpp"
 
 
-// class client {
+class health 
+{
+    private:
+        static uint16_t valid_port(char *port_arg);
+        static bool valid_passord(char *password);
+    public :
+        static uint16_t arg_checker(int ac, char **av);
+};
 
-// };
+class	IMonitor
+{
+	public:
+		virtual	void	push(int fd, short events) = 0;
+		virtual void	pop(int fd) = 0;
+		virtual bool	is_readable(int fd) = 0;
+		virtual bool	is_writeable(int fd) = 0;
+		virtual bool	wait(int timeout = 1) = 0;
+		virtual         ~IMonitor() {};
+};
 
 
-class Server {
+class	Poller : public IMonitor
+{
+	private:
+		std::vector<struct pollfd>	_pfds;
+		std::set<int>		_readable_fds;
+    	std::set<int>		_writeable_fds;
+	public:
+		Poller( void );
+		virtual ~Poller( void );
+		virtual void	push(int fd, short events);
+		virtual void 	pop(int fd);
+		virtual bool	is_readable(int fd);
+		virtual bool	is_writeable(int fd);
+		virtual bool	wait(int timeout = 1);
+};
+
+class Server 
+{
   private:
     unsigned int            _port;
     std::string             _password;
@@ -34,13 +67,8 @@ class Server {
         // void updateFileDescrior(int *ng);
 };
 
-class health {
-    private:
-        static uint16_t valid_port(char *port_arg);
-        static bool valid_passord(char *password);
-    public :
-        static uint16_t arg_checker(int ac, char **av);
-};
+class client {
 
+};
 
 #endif // SERVER_HPP
